Refactor authentication and employee management system
- Changed ClientLoginProjection to use Integer for ID instead of Long.
- Updated AuthService to handle Integer IDs for clients and employees.
- Implemented EmployeeService for employee registration with role and restaurant association.
- Removed unused classes: Client, Employee, and EmployeeRestaurant.
- Introduced new entity classes: ClientEntity, EmployeeEntity, RestaurantEntity, RoleEntity, and WorkScheduleEntity.
- Added EmployeeRegisterRequest DTO for employee registration.
- Created repositories for Employee and Restaurant entities.
- Updated SecurityConfig to restrict access to employee routes.
- Enhanced GlobalExceptionHandler to manage NotFoundException.
- Modified JWT service to work with Integer IDs.
- Updated database initialization script to reflect new entity structure and constraints.
- Removed RestaurantDto as it was no longer needed.

+++ backend_restaurantmanager/src/main/java/com/tfg_rm/backend_restaurantmanager/auth/dto/ClientLoginRequest.java
@@ -22,5 +22,5 @@ public class ClientLoginRequest {
     /**
      * The ID of the restaurant the client wants to access.
      */
-    private Long restaurantId;
+    private Integer restaurantId;
 }
